requirements.txt
datasets~=2.18.0
openai~=1.17.1
backoff~=2.2.1
numpy~=1.26.4
pandas~=2.2.2
python-dotenv~=1.0.1
pytorch~=2.1.2
transformers~=4.36.2
sympy~=1.12
tqdm~=4.66.2
